requirements.txt
streamlit
spacy
nltk
python-docx
PyPDF2
openai
langdetect
indic-transliteration


